add dockerfile for running the app with a web interface
This dockerfile builds on the jlesage/baseimage-gui image that enables
exposing graphical applications as a web application. The idea is to run
the application like usual in a container and expose the GUI via VNC.

The baseimage also includes an NGINX and noVNC installation which means
that no external VNC client is required. Instead, connecting to port
5800 with a browser is enough to use the GUI.

Note that the resulting image is rather large at the moment. Hopefully,
this can be optimized in the future.